FEAT/WIP: new internal download/upload api endpoint
This endpoint is here to download large files out of websocket tunnel to save trafic. Large files sent through websocket tunel are blocking other communication causing timeouts and other errors on both side of WS tunel. This is still work in progress as security and code refactoring needs to be reflected. For now it's just for development and testing purpouses.